Extract code for loading .rten models into a separate module
Extract code from model.rs that deals with loading models in .rten format into
an rten_loader submodule. This is in preparation for adding a separate submodule
which will handle loading from the ONNX format.
